Версия для кластера

+++ Lab_5/PROGRAMM/include/N-body-problem-cuda.cuh
@@ -18,7 +18,7 @@
 
 
 // Используемый тип данных для вычислений
-typedef float mytype;
+typedef double mytype;
